+++ apps/backend/src/database/setup.py
@@ -1,6 +1,6 @@
 from src.database.connection import get_db
 from src.database.collections.at_network_segments import setup_at_network_segments
-from src.database.collections.consumer_units import setup_consumer_units
+from src.database.collections.consumer_units_pj import setup_consumer_units_pj
 from src.database.collections.distribution_indices import setup_distribution_indices
 from src.database.collections.distribution_transformers import setup_distribution_transformers
 from src.database.collections.domain_indicators import setup_domain_indicators
@@ -12,7 +12,7 @@ from src.database.collections.substations import setup_substations
 
 COLLECTION_SETUPS = [
     ("at_network_segments",    setup_at_network_segments),
-    ("consumer_units",         setup_consumer_units),
+    ("consumer_units_pj",      setup_consumer_units_pj),
     ("distribution_indices",   setup_distribution_indices),
     ("distribution_transformers", setup_distribution_transformers),
     ("domain_indicators",      setup_domain_indicators),
@@ -29,7 +29,7 @@ def setup():
 
     for collection_name, setup_fn in COLLECTION_SETUPS:
         if collection_name not in collections_exist:
-            setup_fn(db)
+            setup_fn(db, collections_exist)
 
     return db
